Convert Jenkins interface to Jenkins project interface
Instead of creating and using a "Jenkins Interface"
(sktm.jenkins.skt_jenkins), and then supplying a project name
("jobname") with every method call, create and use a "Jenkins Project
Interface" (sktm.jenkins.Project) and only supply the project name on
its creation.

This simplifes the interface, and removes a bit of code.
This also prepares for abstracting Jenkins away.