patchwork: Add and use ObjectSummary class
Implement patchwork.ObjectSummary class representing a summary of any
mbox-based Patchwork object, storing the object URL, "Date" header, and
patch ID (for patch objects). Switch to storing a list of instances of
that class in patchwork.PatchsetSummary, instead of just URLs, to
represent the patches comprising the patchset.

This avoids having to re-request all the patches in the watcher only to
get their dates and IDs, and prepares the code for the addition of cover
letter support and delayed patches.