Fix bug with pending patch removal
If the patch moved between Patchwork projects during the time it's being
tested, it gets stuck in the pending patch table. Since patch IDs are
unique per instance (and don't depend on the project), change the
DELETE statement to get all known patchsource_ids with given URL and
then remove the appropriate patch from the table. This allows us to
function independently of where the patch moved.

Fixes #57

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>