Refactor patchurl() method
The method does the same thing for both child classes so we can unify
it -- the way to retrieve patch ID is same for both JSON and XMLRPC
objects. Let's also rename the method to fit conventions and use
`os.path.join()` instead of dumb string manipulation since it can handle
slashes.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>