Remove links for FreeDesktop Patchwork fork
We are using the original Patchwork and not the FreeDesktop fork, so any
documentation links should go to the original Patchwork's docs. However,
Patchwork's API docs currently don't contain any of the information that
were supposed to be linked.

Use a comment with explanation based on what Patchwork actually does,
instead of pointing to a documentation of a project we don't interact
with.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>