Make __get_mbox_url_sfx() more reusable
Move the method into PatchworkProject parent class and make it return
the string itself, without any dividers, to make it more reusable. The
caller can then decide if it needs the leading slash or not (most likely
not). This allows us to hide the check for RH fork when using
`get_patch_message()`, but we need to change the method marking to be
able to use it in child classes.

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>