Remove beaker-client from requirements.txt
beaker-client is not used anywhere in the code so it shouldn't be
mentioned in the requirements file. It's only a dependency for Jenkins
to submit jobs and should be marked as such in the README (which it is,
linked as part of skt's dependencies which should be installed in the
prerequisities section).

Signed-off-by: Veronika Kabatova <vkabatov@redhat.com>