Split pending job checking from queueing
Alter the pendingpatches table so that it only contains a timestamp
and a patch_id that links to the id field of the patch table. All
of the relevant data about a patch is in the patch table already and
there is no need to replicate it in pendingpatches.

Change check_pending() so that it searches the pendingjobs table and
takes action on completed baselines and patchwork tests.

Handle any aborted jobs properly without throwing exceptions.

Signed-off-by: Major Hayden <major@redhat.com>