Always update baseline status
When a baseline test is complete, sktm only updates the baseline
record if the status of the commitid got worse (passed before but
failed now). It should always update the status regardless of
the tests improving or getting worse.

Ensure the baseline status is always updated whether it improved or
worsened.

Remove a test for baseline updates that is no longer needed.

Fixes #85.

Signed-off-by: Major Hayden <major@redhat.com>